update today_profit to cumulative_profit

+++ scripts/create_tables.sql
@@ -164,7 +164,7 @@ CREATE TABLE `device_info_daily` (
    `nat_ratio` FLOAT(32) NOT NULL DEFAULT '0',
    `disk_usage` FLOAT(32) NOT NULL DEFAULT '0',
    `upstream_traffic` FLOAT(32) NOT NULL DEFAULT '0',
-   `down_traffic` FLOAT(20) NOT NULL DEFAULT '0',
+   `down_traffic` FLOAT(32) NOT NULL DEFAULT '0',
    `retrieve_count` BIGINT(20) NOT NULL DEFAULT '0',
    PRIMARY KEY USING BTREE (`id`),
    INDEX `idx_device_info_daily_deleted_at` USING BTREE(`deleted_at` ASC)
@@ -187,7 +187,7 @@ CREATE TABLE `device_info_hour` (
  `nat_ratio` FLOAT(32) NOT NULL DEFAULT '0',
  `disk_usage` FLOAT(32) NOT NULL DEFAULT '0',
  `upstream_traffic` FLOAT(32) NOT NULL DEFAULT '0',
- `down_traffic` FLOAT(20) NOT NULL DEFAULT '0',
+ `down_traffic` FLOAT(32) NOT NULL DEFAULT '0',
  `retrieve_count` BIGINT(20) NOT NULL DEFAULT '0',
  PRIMARY KEY USING BTREE (`id`),
  INDEX `idx_device_info_hour_deleted_at` USING BTREE(`deleted_at` ASC)
